Update to read in initial character data
So there's a big issue here with the GrowthRates as a nested class, I can't get the syntax working to be able to access it.  It's probably just because I'm not as familiar with classes in Python but if you can figure that out that'd be great.  I'm pretty sure it reads in the data correctly and saves it correctly, I'm just not sure how to print that part out to be 100% positive.