Slight modifications to calculations
I modified some of the function calls in the calculations and moved that
code out of the Calculations.py file into CharacterData but the Stat
enum is causing syntax problems I don't know how to solve.  I also tried
making character selections dropdowns in the compareCharUi.py file but
the list is horizontal and selects everything (currently commented out).